refactor(utils, documents): Add and integrate getTypeText util for user-friendly document types
